+++ src/compute-engine/solve.ts
@@ -19,49 +19,26 @@ import { expand } from './symbolic/expand';
 
 // Set of rules to find the root(s) for `x`
 export const UNIVARIATE_ROOTS: Rule[] = [
-  // x = 0
-  { match: '_x', replace: 0 },
-
   // ax = 0
-  { match: ['Multiply', '_x', '__a'], replace: 0 },
-
-  // x/a = 0
-  { match: ['Divide', '_x', '_a'], replace: 0 },
-
-  // x/a + b = 0
   {
-    match: ['Add', ['Divide', '_x', '_a'], '__b'],
-    replace: ['Multiply', ['Negate', '_a'], '__b'],
+    match: ['Multiply', '_x', '__a'],
+    replace: 0,
+    id: 'ax',
+    condition: ({ __a }) => !__a.has('_x'),
   },
 
-  // a/x = 0
-  { match: ['Divide', '_a', '_x'], replace: Infinity },
-
-  // x + a = 0
+  // a/x + b = 0
   {
-    match: ['Add', '_x', '__a'],
-    replace: ['Negate', '__a'],
-    id: 'x + a',
+    match: ['Add', ['Divide', '_a', '_x'], '__b'],
+    replace: Infinity,
+    condition: ({ _a, __b }) => !_a.has('_x') && !__b.has('_x'),
   },
-  { match: ['Add', ['Negate', '_x'], '__a'], replace: '__a' },
 
   // ax + b = 0
   {
     match: ['Add', ['Multiply', '_x', '__a'], '__b'],
     replace: ['Divide', ['Negate', '__b'], '__a'],
-  },
-
-  // x^n = 0
-  {
-    match: ['Power', '_x', '_n'],
-    replace: 0,
-  },
-
-  // x^n + c = 0
-  {
-    match: ['Add', ['Power', '_x', '_n'], '__c'],
-    replace: ['Power', ['Negate', '__c'], ['Divide', 1, '_n']],
-    id: 'x^n + c',
+    condition: ({ __a, __b }) => !__a.has('_x') && !__b.has('_x'),
   },
 
   // ax^n + b = 0
@@ -72,11 +49,15 @@ export const UNIVARIATE_ROOTS: Rule[] = [
       ['Power', ['Negate', '__b'], ['Divide', 1, '_n']],
       '_a',
     ],
+    condition: ({ _a, __b, _n }) =>
+      !_a.has('_x') && !__b.has('_x') && !_n.isZero,
   },
 
   //
-  // Quadratic formula (real)
+  // Quadratic formula
   // ax^2 + bx + c = 0
+  //
+
   {
     match: [
       'Add',
@@ -96,7 +77,8 @@ export const UNIVARIATE_ROOTS: Rule[] = [
       ],
       ['Multiply', 2, '__a'],
     ],
-    // (_ce, vars): boolean => vars.x.isReal === true,
+    condition: ({ __a, __b, __c }) =>
+      !__a.has('_x') && !__b.has('_x') && !__c.has('_x'),
   },
 
   {
@@ -118,83 +100,8 @@ export const UNIVARIATE_ROOTS: Rule[] = [
       ],
       ['Multiply', 2, '__a'],
     ],
-    // (_ce, vars): boolean => vars.x.isReal === true,
-  },
-
-  // ax^2 + bx = 0
-  {
-    match: [
-      'Add',
-      ['Multiply', ['Power', '_x', 2], '__a'],
-      ['Multiply', '_x', '__b'],
-    ],
-    replace: 0,
-    // (_ce, vars): boolean => vars.x.isReal === true,
-  },
-  {
-    match: [
-      'Add',
-      ['Multiply', ['Power', '_x', 2], '__a'],
-      ['Multiply', '_x', '__b'],
-    ],
-    replace: ['Divide', ['Negate', '__b'], '__a'],
-    // (_ce, vars): boolean => vars.x.isReal === true,
-  },
-
-  // // ax^2 + c = 0
-  // {
-  //   match: ['Add', ['Multiply', ['Power', '_x', 2], '__a'], '__c'],
-  //   replace: ['Sqrt', ['Divide', ['Negate', '__c'], '__a']],
-  //   // (_ce, vars): boolean => vars.x.isReal === true,
-  // },
-  // {
-  //   match: ['Add', ['Multiply', ['Power', '_x', 2], '__a'], '__b'],
-  //   replace: ['Negate', ['Sqrt', ['Divide', ['Negate', '__b'], '__a']]],
-  //   // (_ce, vars): boolean => vars.x.isReal === true,
-  // },
-
-  // x^2 + bx + c = 0
-  {
-    match: ['Add', ['Power', '_x', 2], ['Multiply', '__b', '_x'], '__c'],
-    replace: [
-      'Divide',
-      [
-        'Add',
-        ['Negate', '__b'],
-        ['Sqrt', ['Subtract', ['Square', '__b'], ['Multiply', 4, '__c']]],
-      ],
-      2,
-    ],
-    // (_ce, vars): boolean => vars.x.isReal === true,
-  },
-
-  {
-    match: ['Add', ['Power', '_x', 2], ['Multiply', '__b', '_x'], '__c'],
-    replace: [
-      'Divide',
-      [
-        'Subtract',
-        ['Negate', '__b'],
-        ['Sqrt', ['Subtract', ['Square', '__b'], ['Multiply', 4, '__c']]],
-      ],
-      2,
-    ],
-    // (_ce, vars): boolean => vars.x.isReal === true,
-  },
-
-  // x^2 + bx = 0
-  {
-    match: ['Add', ['Power', '_x', 2], ['Multiply', '__b', '_x']],
-    replace: 0,
-    // (_ce, vars): boolean => vars.x.isReal === true,
-    id: 'x^2 + bx -> 0',
-  },
-
-  {
-    match: ['Add', ['Power', '_x', 2], ['Multiply', '__b', '_x']],
-    replace: ['Negate', '__b'],
-    // (_ce, vars): boolean => vars.x.isReal === true,
-    id: 'x^2 + bx -> -b',
+    condition: ({ __a, __b, __c }) =>
+      !__a.has('_x') && !__b.has('_x') && !__c.has('_x'),
   },
 
   // a * e^(bx) + c = 0
@@ -206,7 +113,9 @@ export const UNIVARIATE_ROOTS: Rule[] = [
     ],
     replace: ['Divide', ['Ln', ['Negate', ['Divide', '__c', '__a']]], '__b'],
     condition: ({ __a, __c }, ce) =>
-      (!__a.isZero && ce.div(__c, __a).isNegative) ?? false,
+      ((!__a.isZero && ce.div(__c, __a).isNegative) ?? false) &&
+      !__a.has('_x') &&
+      !__c.has('_x'),
   },
 
   // a * e^(x) + c = 0
@@ -214,22 +123,24 @@ export const UNIVARIATE_ROOTS: Rule[] = [
     match: ['Add', ['Multiply', '__a', ['Exp', '_x']], '__c'],
     replace: ['Ln', ['Negate', ['Divide', '__c', '__a']]],
     condition: ({ __a, __c }, ce) =>
-      (!__a.isZero && ce.div(__c, __a).isNegative) ?? false,
+      ((!__a.isZero && ce.div(__c, __a).isNegative) ?? false) &&
+      !__a.has('_x') &&
+      !__c.has('_x'),
   },
 
-  // e^(x) + c = 0
-  {
-    match: ['Add', ['Exp', '_x'], '__c'],
-    replace: ['Ln', ['Negate', '__c']],
-    condition: ({ __c }) => __c.isNegative ?? false,
-  },
+  // // e^(x) + c = 0
+  // {
+  //   match: ['Add', ['Exp', '_x'], '__c'],
+  //   replace: ['Ln', ['Negate', '__c']],
+  //   condition: ({ __c }) => __c.isNegative ?? false,
+  // },
 
-  // e^(bx) + c = 0
-  {
-    match: ['Add', ['Exp', ['Multiply', '__b', '_x']], '__c'],
-    replace: ['Divide', ['Ln', ['Negate', '__c']], '__b'],
-    condition: ({ __c }) => __c.isNegative ?? false,
-  },
+  // // e^(bx) + c = 0
+  // {
+  //   match: ['Add', ['Exp', ['Multiply', '__b', '_x']], '__c'],
+  //   replace: ['Divide', ['Ln', ['Negate', '__c']], '__b'],
+  //   condition: ({ __c }) => __c.isNegative ?? false,
+  // },
 
   // // a * log_b(x) + c = 0
   // {
@@ -245,38 +156,15 @@ export const UNIVARIATE_ROOTS: Rule[] = [
   //   condition: ({ __a, __b }) => (!__a.isZero && __b.isPositive) ?? false,
   // },
 
-  // |ax + b| + c = 0
-  {
-    match: ['Add', ['Abs', ['Add', ['Multiply', '__a', '_x'], '__b']], '__c'],
-    replace: ['Divide', ['Subtract', '__b', '__c'], '__a'],
-  },
-  {
-    match: ['Add', ['Abs', ['Add', ['Multiply', '__a', '_x'], '__b']], '__c'],
-    replace: ['Divide', ['Negate', ['Add', '__b', '__c'], '__a']],
-  },
-
-  // // x^2 + c = 0
+  // // |ax + b| + c = 0
   // {
-  //   match: ['Add', ['Power', '_x', 2], '__c'],
-  //   replace: ['Power', ['Negate', '__c'], 'Half'],
-  //   // (_ce, vars): boolean => vars.x.isReal === true,
+  //   match: ['Add', ['Abs', ['Add', ['Multiply', '__a', '_x'], '__b']], '__c'],
+  //   replace: ['Divide', ['Subtract', '__b', '__c'], '__a'],
   // },
-
   // {
-  //   match: ['Add', ['Power', '_x', 2], '__c'],
-  //   replace: ['Negate', ['Power', ['Negate', '__c'], 'Half']],
-  //   // (_ce, vars): boolean => vars.x.isReal === true,
+  //   match: ['Add', ['Abs', ['Add', ['Multiply', '__a', '_x'], '__b']], '__c'],
+  //   replace: ['Divide', ['Negate', ['Add', '__b', '__c'], '__a']],
   // },
-
-  // Quadratic formula (complex)
-  // [
-  //   '$ax^2 + bx + c$',
-  //   [
-  //     '$-\\frac{b}{2a} - \\imaginaryI \\frac{\\sqrt{4ac - b^2}}{2a}$',
-  //     '$-\\frac{b}{2a} + \\imaginaryI \\frac{\\sqrt{4ac - b^2}}{2a}$',
-  //   ],
-  //   (_ce, vars): boolean => vars.x.isImaginary === true,
-  // ],
 ];
 
 /**
@@ -297,21 +185,39 @@ export function findUnivariateRoots(
   const rules = ce.cache('univariate-roots-rules', () =>
     boxRules(ce, UNIVARIATE_ROOTS)
   );
-  let result = matchRules(
-    expr.subs({ [x]: '_x' }, { canonical: false }),
-    rules,
-    { _x: ce.symbol('_x') }
+  // Make the unknown '_x' so that we can match against it
+  let exprs = [expr.subs({ [x]: '_x' }, { canonical: false })];
+
+  let result = exprs.flatMap((expr) =>
+    matchRules(expr, rules, { _x: ce.symbol('_x') })
   );
 
+  // If we didn't find a solution yet, try modifying the expression
+  //
+  // Note: @todo we can try different heuristics here:
+  // Collection: reduce the numbers of occurences of the unknown
+  // Attraction: bring the occurences of the unknonw closer together
+  // Function Swapping: replacing function with ones easier to solve
+  //    - square roots: square both sides
+  //    - logs: exponentiate both sides
+  //    - trig functions: use inverse trig functions
+  // Homogenization: replace a function of the unknown by a new variable,
+  // e.g. exp(x) -> y, then solve for y
+
   if (result.length === 0) {
-    const expandedExpr = expand(expr.canonical);
-    if (expandedExpr === null) return [];
-    result = matchRules(
-      expandedExpr.subs({ [x]: '_x' }, { canonical: false }),
-      rules,
-      {
-        _x: ce.symbol('_x'),
-      }
+    exprs = exprs.flatMap((expr) => harmonize(expr));
+    result = exprs.flatMap((expr) =>
+      matchRules(expr, rules, { _x: ce.symbol(x) })
+    );
+  }
+
+  if (result.length === 0) {
+    exprs = exprs
+      .flatMap((expr) => expand(expr.canonical))
+      .filter((x) => x !== null) as BoxedExpression[];
+    exprs = exprs.flatMap((expr) => harmonize(expr));
+    result = exprs.flatMap((expr) =>
+      matchRules(expr, rules, { _x: ce.symbol(x) })
     );
   }
 
@@ -331,8 +237,9 @@ export function univariateSolve(
 ): SemiBoxedExpression[] | null {
   const ce = expr.engine;
   const name = expr.head;
-  if (name === 'And') {
+  if (name === 'Tuple') {
     // @todo: System of equations
+    return null;
   }
 
   if (
@@ -343,11 +250,112 @@ export function univariateSolve(
     return null;
   }
 
-  const rhs = expr.op2;
   let lhs: SemiBoxedExpression = expr.op1;
+  const rhs = expr.op2;
   if (rhs.isNotZero === true) lhs = ['Subtract', lhs, rhs];
 
   const roots = findUnivariateRoots(ce.box(lhs), x);
   if (roots.length === 0) return null;
   return roots;
+}
+
+export const HARMONIZATION_RULES: Rule[] = [
+  // |ax + b| + c -> ax+b+c, -ax-b+c
+  {
+    match: ['Add', ['Abs', ['Add', ['Multiply', '__a', '_x'], '__b']], '__c'],
+    replace: ['Add', ['Multiply', '__a', '_x'], '__b', '__c'],
+  },
+  {
+    match: ['Add', ['Abs', ['Add', ['Multiply', '__a', '_x'], '__b']], '__c'],
+    replace: [
+      'Add',
+      ['Negate', ['Multiply', '__a', '_x']],
+      ['Negate', '__b'],
+      '__c',
+    ],
+  },
+  // a(b^n) -> a
+  {
+    match: ['Multiply', '__a', ['Power', '_b', '_n']],
+    replace: '_b',
+    condition: ({ __a, _b, _n }) =>
+      !__a.has('_x') && _b.has('_x') && !_n.isZero,
+  },
+  // a√b(x)  -> a^2 b(x)
+  {
+    match: ['Multiply', '__a', ['Sqrt', '_b']],
+    replace: ['Multiply', ['Square', '_a'], '__b'],
+    condition: ({ _b }) => _b.has('_x'),
+  },
+  // a(x)/b -> a(x)
+  {
+    match: ['Divide', '_a', '_b'],
+    replace: '_a',
+    condition: ({ _a, _b }) => _a.has('_x') && !_b.isZero,
+  },
+  // ab(x) -> b(x)
+  {
+    match: ['Multiply', '__a', '_b'],
+    replace: '_b',
+    condition: ({ __a, _b }) => !__a.has('_x') && _b.has('_x'),
+  },
+  // ln(a(x))+ln(b(x))+c -> ln(a(x)b(x)) + c
+  {
+    match: ['Add', ['Ln', '_a'], ['Ln', '_b'], '__c'],
+    replace: ['Add', ['Ln', ['Multiply', '_a', '_b']], '__c'],
+  },
+  // e^a * e^b -> e^(a+b)
+  {
+    match: ['Multiply', ['Exp', '__a'], ['Exp', '__b'], '__c'],
+    replace: ['Multiply', ['Exp', ['Add', '_a', '_b']], '__c'],
+  },
+  // ln(f(x)) -> f(x) - 1
+  {
+    match: ['Ln', '_a'],
+    replace: ['Subtract', '_a', 1],
+    condition: ({ _a }) => _a.has('_x'),
+  },
+  // sin(f(x)) -> f(x)
+  {
+    match: ['Sin', '_a'],
+    replace: '_a',
+    condition: ({ _a }) => _a.has('_x'),
+  },
+  // cos(f(x)) -> f(x) - π/2
+  {
+    match: ['Cos', '_a'],
+    replace: ['Subtract', '_a', ['Divide', 'Pi', 2]],
+    condition: ({ _a }) => _a.has('_x'),
+  },
+  // tan(f(x)) -> f(x) - π/4
+  {
+    match: ['Tan', '_a'],
+    replace: '_a',
+    condition: ({ _a }) => _a.has('_x'),
+  },
+  // sin(a) + cos(a) -> 1
+  {
+    match: ['Add', ['Sin', '_a'], ['Cos', '_a']],
+    replace: 1,
+    condition: ({ _a }) => _a.has('_x'),
+  },
+  // sin^2(a) - cos^2(a) -> sin(x) +/- √(2)/2
+  {
+    match: ['Subtract', ['Square', ['Sin', '_a']], ['Square', ['Cos', '_a']]],
+    replace: ['PlusMinus', ['Sin', '_a'], ['Divide', ['Sqrt', 2], 2]],
+    condition: ({ _a }) => _a.has('_x'),
+  },
+];
+
+/** Transform expr into one or more equivalent expressions that
+ * are easier to solve
+ */
+function harmonize(expr: BoxedExpression): BoxedExpression[] {
+  const ce = expr.engine;
+  const rules = ce.cache('univariate-roots-rules', () =>
+    boxRules(ce, HARMONIZATION_RULES)
+  );
+  let result = matchRules(expr, rules, { _x: ce.symbol('_x') });
+
+  return result;
 }