+++ test/compute-engine/solve.test.ts
@@ -4,6 +4,50 @@ function expr(s: string) {
   return engine.parse(s);
 }
 
+const eqn = engine.box(['Multiply', '5', 'x']);
+
+console.log(engine.pattern(['Multiply', 'x', '__a']).match(eqn));
+
+describe('SOLVING A QUADRATIC EQUATION', () => {
+  const ce = engine;
+
+  it('should solve bx', () => {
+    const eqn = ce.box(['Multiply', 5, 'x']);
+    const result = eqn.solve('x')?.map((x) => x.json);
+    expect(result).toMatchInlineSnapshot(`[0]`);
+  });
+
+  it('should solve bx + c', () => {
+    const eqn = ce.box(['Add', ['Multiply', 5, 'x'], -10]);
+    const result = eqn.solve('x')?.map((x) => x.json);
+    expect(result).toMatchInlineSnapshot(`[2]`);
+  });
+
+  it('should solve ax^2', () => {
+    const eqn = ce.box(['Multiply', 16, ['Square', 'x']]);
+    const result = eqn.solve('x')?.map((x) => x.json);
+    expect(result).toMatchInlineSnapshot(`[0]`);
+  });
+
+  it('should solve ax^2 + c', () => {
+    const eqn = ce.box(['Add', ['Multiply', 2, ['Square', 'x']], -16]);
+    const result = eqn.solve('x')?.map((x) => x.json);
+    expect(result).toMatchInlineSnapshot(`[2]`);
+  });
+
+  it('should solve ax^2 + bx + c', () => {
+    const eqn = ce.box([
+      'Add',
+      ['Multiply', 2, ['Square', 'x']],
+      ['Multiply', 6, 'x'],
+      4,
+    ]);
+
+    const result = eqn.solve('x')?.map((x) => x.json);
+    expect(result).toMatchInlineSnapshot(`[-1, -2]`);
+  });
+});
+
 describe('expr.solve()', () => {
   test('should solve an assignment', () => {
     const e = expr('x = 5');
@@ -35,12 +79,6 @@ describe('expr.solve()', () => {
     expect(result).toEqual([3]);
   });
 
-  test('should solve a quadratic equation', () => {
-    const e = expr('x^2 + 3x + 2 = 0');
-    const result = e.solve('x')?.map((x) => x.json);
-    expect(result).toEqual([-1, -2]);
-  });
-
   test('should solve an equation with a fractional coefficient', () => {
     const e = expr('\\frac{2}{3}x + \\frac{1}{3} = 5');
     const result = e.solve('x')?.map((x) => x.json);
@@ -60,6 +98,12 @@ describe('expr.solve()', () => {
     const e = expr('x^2 + 1 = 0');
     const result = e.solve('x')?.map((x) => x.json);
     expect(result).toEqual([['Complex', 0, 1]]);
+  });
+
+  test('should **NOT** solve a quasi-quadratic equation', () => {
+    const e = expr('x^2 + 3x + 2 + \\sin(x) = 0');
+    const result = e.solve('x')?.map((x) => x.json);
+    expect(result).toMatchInlineSnapshot(`[]`);
   });
 
   // test('should solve an inequality', () => {